refactor: file organization
